+++ gwcelery/conf/__init__.py
@@ -85,11 +85,13 @@ preliminary_alert_far_threshold = {'cbc': 1 / (30 * 86400),
 """Group specific maximum false alarm rate to consider
 sending preliminary alerts."""
 
-preliminary_alert_trials_factor = dict(cbc=4.0, burst=2.0)
+preliminary_alert_trials_factor = dict(cbc=5.0, burst=5.0)
 """Trials factor corresponding to trigger categories.
-The trials factor is simply the number of pipelines running for each
-type of search. The value corresponds to pipelines gstlal, pycbc,
-mbtaonline, spiir for CBC and cwb and olib for Burst."""
+For CBC, trials factor is the number of pipelines plus the external
+coincidence search. For Burst, this is the total number of searches
+plus the external coincidence search.
+CBC pipelines are gstlal, pycbc, mbtaonline, spiir.
+Burst searches are cwb.allsky, cwb.bbh, cwb.imbh and olib.allsky."""
 
 orchestrator_timeout = 15.0
 """The orchestrator will wait this many seconds from the time of the