+++ CHANGES.md
@@ -13,6 +13,10 @@
 -   The Nagios monitoring plugin will now report on the status of LVAlert
     subscriptions.
 
+-   Change trails factor to 5 for both CBC and Burst categories. CBC includes
+    the 4 CBC pipelines. Burst includes the 4 searches performed in total by
+    the 2 Burst pipelines. An additional external coincidence search.
+
 ## 0.2.4 (2018-12-17)
 
 -   Fix broken links in log messages due to changes in GraceDb URL routes.